Change ids to long

+++ src/main/java/edu/mcw/scge/dao/implementation/EditorDao.java
@@ -30,7 +30,7 @@ public class EditorDao extends AbstractDAO {
         return execute(q, personId);
     }
 
-    public List<Editor> getEditorById(int id) throws Exception {
+    public List<Editor> getEditorById(long id) throws Exception {
         String sql="select * from editor where editor_id=?";
         EditorQuery q=new EditorQuery(this.getDataSource(), sql);
         return execute(q, id);
@@ -41,14 +41,14 @@ public class EditorDao extends AbstractDAO {
         EditorQuery q=new EditorQuery(this.getDataSource(), sql);
         return execute(q, guideId);
 	}
-    public int insertEditor(Editor editor) throws Exception{
+    public long insertEditor(Editor editor) throws Exception{
 
         String sql = "insert into editor ( editor_id, subtype, species, pam_preference, editor_variant, \n" +
                 "fusion, activity, dsb_cleavage_type, target_sequence, source, \n" +
                 "type, symbol, alias,  substrate_target, protein_sequence, editor_description, annotated_map,tier ) " +
                 "values (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
 
-        int editorId = this.getNextKeyFromSequence("editor_seq");
+        long editorId = this.getNextKeyFromSequence("editor_seq");
 
 
         update(sql, editorId,editor.getSubType(),editor.getSpecies(),editor.getPamPreference(),
@@ -60,7 +60,7 @@ public class EditorDao extends AbstractDAO {
         return editorId;
     }
 
-    public int getEditorId(Editor editor) throws Exception {
+    public long getEditorId(Editor editor) throws Exception {
 
         String sql = "select * from editor where species =? and type=? and subtype=? and symbol = ?";
 
@@ -69,7 +69,7 @@ public class EditorDao extends AbstractDAO {
     }
 
 
-    public boolean verifyEditorAccess(int editorId, int personId) throws Exception {
+    public boolean verifyEditorAccess(long editorId, int personId) throws Exception {
         String sql="(select e.* from editor e left outer join experiment_record r on e.editor_id=r.editor_id\n" +
                 "left outer join experiment x on x.experiment_id=r.experiment_id " +
                 "left outer join study s on s.study_id =x.study_id " +
@@ -81,7 +81,7 @@ public class EditorDao extends AbstractDAO {
         List<Editor> editorList= execute(q, personId, editorId, editorId);
         return editorList.size() > 0;
     }
-    public void updateEditorTier(int tier, int editorId) throws Exception{
+    public void updateEditorTier(int tier, long editorId) throws Exception{
         String sql="update editor set tier=? where editor_id=?";
         update(sql, tier, editorId);
     }
