+++ src/main/java/edu/mcw/scge/dao/implementation/ExperimentDao.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 public class ExperimentDao extends AbstractDAO {
 
-    public List<String> getExperimentRecordTissueList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordTissueList(long experimentId) throws Exception {
         String sql="select distinct ot.term from experiment_record ex " +
                 "inner join ont_terms ot on ex.tissue_id=ot.term_acc " +
                 "where ex.experiment_id=? order by ot.term";
@@ -21,7 +21,7 @@ public class ExperimentDao extends AbstractDAO {
         return execute(q, experimentId);
     }
 
-    public List<String> getExperimentRecordCellTypeList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordCellTypeList(long experimentId) throws Exception {
         String sql="select distinct ot.term from experiment_record ex " +
                 "inner join ont_terms ot on ex.cell_type=ot.term_acc " +
                 "where ex.experiment_id=? order by ot.term";
@@ -30,7 +30,7 @@ public class ExperimentDao extends AbstractDAO {
         return execute(q, experimentId);
     }
 
-    public List<String> getExperimentRecordConditionList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordConditionList(long experimentId) throws Exception {
         String sql="select distinct ex.name from experiment_record ex " +
                 " where ex.experiment_id=? order by name";
 
@@ -38,7 +38,7 @@ public class ExperimentDao extends AbstractDAO {
         return execute(q, experimentId);
     }
 
-    public List<String> getExperimentRecordEditorList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordEditorList(long experimentId) throws Exception {
         String sql="select distinct e.symbol from experiment_record ex " +
                 " inner join experiment x on x.experiment_id=ex.experiment_id " +
                 " inner join editor e on ex.editor_id = e.editor_id " +
@@ -48,21 +48,21 @@ public class ExperimentDao extends AbstractDAO {
         return execute(q, experimentId);
     }
 
-    public List<String> getExperimentRecordVectorList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordVectorList(long experimentId) throws Exception {
         String sql="select distinct v.name from vector v, experiment e, experiment_record er, vector_associations va " +
                 "     where e.experiment_id=? and e.experiment_id=er.experiment_id and va.vector_id=v.vector_id and va.experiment_record_id=er.experiment_record_id order by v.name";
         StringListQuery q=new StringListQuery(this.getDataSource(), sql);
         return execute(q, experimentId);
     }
 
-    public List<String> getExperimentRecordGuideList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordGuideList(long experimentId) throws Exception {
         String sql="select distinct g.guide from guide g, experiment e, experiment_record er, guide_associations ga " +
                 "     where e.experiment_id=? and e.experiment_id=er.experiment_id and ga.guide_id=g.guide_id and ga.experiment_record_id=er.experiment_record_id order by g.guide ";
         StringListQuery q=new StringListQuery(this.getDataSource(), sql);
         return execute(q, experimentId);
     }
 
-    public List<String> getExperimentRecordModelList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordModelList(long experimentId) throws Exception {
         String sql="select distinct m.name from experiment_record ex " +
                 " inner join model m on ex.model_id = m.model_id " +
                 " where ex.experiment_id=? order by m.name";
@@ -70,7 +70,7 @@ public class ExperimentDao extends AbstractDAO {
         StringListQuery q=new StringListQuery(this.getDataSource(), sql);
         return execute(q, experimentId);
     }
-    public List<String> getExperimentRecordDeliverySystemList(int experimentId) throws Exception {
+    public List<String> getExperimentRecordDeliverySystemList(long experimentId) throws Exception {
         String sql="select distinct d.ds_name from experiment_record ex " +
                 "inner join delivery_system d on ex.ds_id = d.ds_id " +
                 "where ex.experiment_id=? order by d.ds_name";
@@ -82,7 +82,7 @@ public class ExperimentDao extends AbstractDAO {
         return returnVal;
     }
 
-    public List<ExperimentRecord> getExperimentRecords(int experimentId) throws Exception {
+    public List<ExperimentRecord> getExperimentRecords(long experimentId) throws Exception {
         String sql="select ex.*, e.symbol, d.ds_type, d.ds_name, m.name as modelName, x.type, ot.term, ct.term as cellTerm  from experiment_record ex " +
                 " left outer join experiment x on x.experiment_id=ex.experiment_id " +
                 "left outer join editor e on ex.editor_id = e.editor_id " +
@@ -127,7 +127,7 @@ public class ExperimentDao extends AbstractDAO {
         ExperimentQuery q=new ExperimentQuery(this.getDataSource(), sql);
         return execute(q, studyId, personId);
     }
-    public Experiment getExperiment(int experimentId) throws Exception {
+    public Experiment getExperiment(long experimentId) throws Exception {
         String sql="select ex.* from experiment ex " +
 
                 "where ex.experiment_id=?";
@@ -137,7 +137,7 @@ public class ExperimentDao extends AbstractDAO {
 
     }
 
-    public List<ExperimentRecord> getExperimentsByEditor(int editorId) throws Exception {
+    public List<ExperimentRecord> getExperimentsByEditor(long editorId) throws Exception {
         String sql="select s.study, ex.*, e.symbol, d.ds_type, d.ds_name, m.name as modelName, x.type, ot.term, ct.term as cellTerm from experiment_record ex " +
                 " left outer join experiment x on x.experiment_id=ex.experiment_id " +
                 "left outer join editor e on ex.editor_id = e.editor_id " +
@@ -153,7 +153,7 @@ public class ExperimentDao extends AbstractDAO {
         return execute(q, editorId);
     }
 
-    public List<ExperimentRecord> getExperimentsByModel(int modelId) throws Exception {
+    public List<ExperimentRecord> getExperimentsByModel(long modelId) throws Exception {
         String sql="select s.study, ex.*, e.symbol, d.ds_type, d.ds_name, m.name as modelName, x.type, ot.term, ct.term as cellTerm from experiment_record ex " +
                 " left outer join experiment x on x.experiment_id=ex.experiment_id " +
                 "left outer join editor e on ex.editor_id = e.editor_id " +
@@ -169,7 +169,7 @@ public class ExperimentDao extends AbstractDAO {
         return execute(q, modelId);
     }
 
-    public List<ExperimentRecord> getExperimentsByDeliverySystem(int dsId) throws Exception {
+    public List<ExperimentRecord> getExperimentsByDeliverySystem(long dsId) throws Exception {
         String sql="select s.study, ex.*, e.symbol, d.ds_type, d.ds_name, m.name as modelName, x.type, ot.term, ct.term as cellTerm from experiment_record ex " +
                 " left outer join experiment x on x.experiment_id=ex.experiment_id " +
                 "left outer join editor e on ex.editor_id = e.editor_id " +
@@ -227,7 +227,7 @@ public class ExperimentDao extends AbstractDAO {
         String sql = "insert into experiment (experiment_id,name,study_id, " +
                 "type  ) values (?,?,?,?)";
 
-        int experimentId = this.getNextKeyFromSequence("experiment_seq");
+        long experimentId = this.getNextKeyFromSequence("experiment_seq");
         
         update(sql, experimentId,experiment.getName(),experiment.getStudyId(),experiment.getType());
     }
