+++ src/main/java/edu/mcw/scge/dao/implementation/ExperimentResultDao.java
@@ -8,47 +8,47 @@ import edu.mcw.scge.datamodel.ExperimentResultDetail;
 import java.util.List;
 
 public class ExperimentResultDao extends AbstractDAO {
-    public List<ExperimentResultDetail> getResultsByExperimentRecId(int expRecId) throws Exception {
+    public List<ExperimentResultDetail> getResultsByExperimentRecId(long expRecId) throws Exception {
         String sql="select r.*,d.replicate,d.result from experiment_result r inner join experiment_result_detail d " +
                 "on r.result_id = d.result_id where r.experiment_record_id=? ";
         ExperimentResultDetailQuery q=new ExperimentResultDetailQuery(this.getDataSource(), sql);
         return execute(q, expRecId);
     }
 
-    public List<ExperimentResultDetail> getResultsByExperimentId(int expId) throws Exception {
+    public List<ExperimentResultDetail> getResultsByExperimentId(long expId) throws Exception {
         String sql="select r.*,d.replicate,d.result from experiment_result r inner join experiment_result_detail d " +
                 "on r.result_id = d.result_id inner join experiment_record e on r.experiment_record_id = e.experiment_record_id" +
                 " where e.experiment_id=? ";
         ExperimentResultDetailQuery q=new ExperimentResultDetailQuery(this.getDataSource(), sql);
         return execute(q, expId);
     }
-    public List<ExperimentResultDetail> getResultsByExpResType(int expRecId,String resultType) throws Exception {
+    public List<ExperimentResultDetail> getResultsByExpResType(long expRecId,String resultType) throws Exception {
         String sql="select r.*,d.replicate,d.result from experiment_result r inner join experiment_result_detail d " +
                 "on r.result_id = d.result_id where r.experiment_record_id=? and r.result_type = ?";
         ExperimentResultDetailQuery q=new ExperimentResultDetailQuery(this.getDataSource(), sql);
         return execute(q, expRecId,resultType);
     }
 
-    public List<String> getResTypeByExpId(int expId) throws Exception {
+    public List<String> getResTypeByExpId(long expId) throws Exception {
         String sql="select distinct(r.result_type) from experiment_result r inner join experiment_record e " +
                 "on e.experiment_record_id = r.experiment_record_id where e.experiment_id = ?";
         StringListQuery q=new StringListQuery(this.getDataSource(), sql);
         return execute(q, expId);
     }
-    public List<String> getUnitsByExpId(int expId) throws Exception {
+    public List<String> getUnitsByExpId(long expId) throws Exception {
         String sql="select distinct(r.units) from experiment_result r inner join experiment_record e " +
                 "on e.experiment_record_id = r.experiment_record_id where e.experiment_id = ?";
         StringListQuery q=new StringListQuery(this.getDataSource(), sql);
         return execute(q, expId);
     }
 
-	public int insertExperimentResult(ExperimentResultDetail expResult) throws Exception{
+	public long insertExperimentResult(ExperimentResultDetail expResult) throws Exception{
 
         String sql = "insert into experiment_result (result_id, number_of_samples, units," +
                 " assay_description, experiment_record_id, result_type,edit_type  )" +
                 " values (?,?,?,?,?,?,?)";
-        
-        int resultId = this.getNextKeyFromSequence("result_seq");
+
+        long resultId = this.getNextKeyFromSequence("result_seq");
         
         update(sql, resultId,
                 expResult.getNumberOfSamples(),expResult.getUnits(),expResult.getAssayDescription(),
