+++ src/main/java/edu/mcw/scge/dao/spring/ExperimentResultDetailQuery.java
@@ -15,8 +15,8 @@ public class ExperimentResultDetailQuery extends MappingSqlQuery<ExperimentResul
     protected ExperimentResultDetail mapRow(ResultSet rs, int rowNum) throws SQLException {
 
         ExperimentResultDetail e= new ExperimentResultDetail();
-        e.setResultId(rs.getInt("result_id"));
-        e.setExperimentRecordId(rs.getInt("experiment_record_id"));
+        e.setResultId(rs.getLong("result_id"));
+        e.setExperimentRecordId(rs.getLong("experiment_record_id"));
         e.setAssayDescription(rs.getString("assay_description"));
         e.setNumberOfSamples(rs.getInt("number_of_samples"));
         e.setUnits(rs.getString("units"));
